created utils.py for orginization
